% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printkv_function.R
\name{printkv}
\alias{printkv}
\title{printkv}
\usage{
printkv(key, val, sep = " = ", fill = 0, fill_char = " ", round = FALSE)
}
\arguments{
\item{key}{the content of the key}

\item{val}{content of the value}

\item{sep}{the character used to separate a key from a value.
defaults to ": "}

\item{fill}{integer. The minimum number of columns to fill up before the
separator is displayed. Use this to line values up.}

\item{round}{used for setting the number of decimal places to show (if val
       was a numeric type).
       If round is FALSE (Default), then no rounding occurs.
       If round is an integer, then val is rounded to that many decimal
       places}
}
\description{
Print a key value pair. You can specify that the value be positioned at a
certain minimum column number, which can be used to line up multiple
prinkv() function calls.
You can also specity how the gap between the key and the value should be
filled up, aswell as the separator.
}
\examples{
printkv("age", 34, sep=": ", fill=10, fill_char=".")
printkv("age", 56)
}
\keyword{ascii,}
\keyword{fancyprint,}
\keyword{print}

