% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printkv_function.R
\name{printkv}
\alias{printkv}
\title{printkv}
\usage{
printkv(key, val, sep = ": ", fill = 0, fill_char = " ", round = FALSE,
  vas = TRUE)
}
\arguments{
\item{key}{the content of the key}

\item{val}{content of the value}

\item{sep}{the character used to separate a key from a value.
defaults to ": "}

\item{fill}{integer. The minimum number of columns to fill up before the
separator is displayed. Use this to line values up.}

\item{round}{used for setting the number of decimal places to show (if val
was a numeric type).
If round is FALSE (Default), then no rounding occurs.
If round is an integer, then val is rounded to that many decimal
places}

\item{vas}{(boolean) Vectors As Strings. If a vector is fed as the val
       argument, should it convert the entire vector into a string? This will
       print all the values of the vector as a string if TRUE is used.
       FALSE = print a key value pair for each element of the vector.
       See the examples section to see how this is used.
       (DEFAULT = TRUE)}
}
\description{
Print a key value pair. You can specify that the value be positioned at a
certain minimum column number, which can be used to line up multiple
prinkv() function calls.
You can also specity how the gap between the key and the value should be
filled up, aswell as the separator.
}
\examples{
printkv("name", "Joe", sep="= ", fill=10, fill_char=".")
printkv("age", 34, sep="= ", fill=10, fill_char=".")
printkv("age", 56)

printkv("rank", c(1,2,3,4))
[1] "rank: (1,2,3,4)"

printkv("rank", c(1,2,3,4), vas=FALSE)
[1] "rank: 1" "rank: 2" "rank: 3" "rank: 4"
}
\keyword{ascii,}
\keyword{fancyprint,}
\keyword{print}

